Use portions of LLVM (assembler, strip) and GNU gold (#1)
GNU Binutils have problems dealing with diacritic characters on Windows platform.
Fixing the issue in the Binutils codebase is not straightforward, so we decided to
use portions of the LLVM toolchain (`llvm-mc`, the assembler, and `llvm-objcopy` 
which doubles as `strip`) instead, since they don't have said problem with diacritcs.

The initial plan was to switch the linker to LLVM's `lld` as well, but since its binary
comes up at 70mb (twice that on mac, with two architectures) we decided to keep
using GNU ld (`gold`) for the moment. 

In order to maintain compatibility with Mono AOT and Xamarin.Android, it was
necessary to develop a GNU Assembler (`gas`) wrapper around `llvm-mc`.

Implement build and packaging for LLVM, GNU Binutils and XA own utilities.