[WIP] Disable building arm64 binaries on macOS
Something's not right with the environment on CI, the arm64 builds
fail to find `math.h` when included from `cmath`.  It appears there's a
mixup between Xcode.app and Xcode command line tools (it **seems** the
former doesn't have everything installed, while the latter has, but is
not completely compatible with the former). Disable for now.