[ci] Migrate to the 1ES template (#8)
Context: https://aka.ms/1espt

The build pipeline has been updated to extend the 1ES pipeline template,
which will keep the pipeline up to date with the latest compliance and
security requirements.

Compliance tasks and scans will run automatically as part of artifact
upload steps, which are now referred to as "outputs".  Template outputs
have replaced all instances of the `PublishPipelineArtifact` task.

The new compliance steps appear to have added ~4 hours to the
build in the worst-case scenario (from 1.5 hours to 5.5 hours).  This
appears to mostly be a result of CodeQL.