Build `as.exe` on windows (#10)
Build our `GNU as` wrapper around LLVM's `llvm-mc` assembler entirely on Windows, using MSVC.

This is required to satisfy the API Scan requirements for the binary.

As part of this, the wrapper transitioned from using the GNU `getopt` command-line parsing API to a header-only
portable C++ library, `cxxopt`.  The move was necessary because the MSVC toolchain lacks an implementation of
`getopt` (while it is part of the MinGW's POSIX compatibility layer, and previously the utility was built with MinGW
on Linux).

Additionally, the Linux hosts now use g++ 13 to build the utilities.  Pipeline error handling has been fixed, where
it would previously ignore build errors leading to false positive "green" builds.