Fix `as.exe` crash on Windows (#20)
Rewrote the GNU `as` wrapper to make it work properly when compiled by MSVC.
The main problem was that handling of command line argument stopped working
after switching compilers.  Before that, the wrapper was built using MinGW which used 
a POSIX compatibility mode, and that allowed us to use nearly exactly the same code on
Unix as on Windows.  

However, after switching to MSVC, we encountered a problem where
command line arguments which had UTF/Unicode characters stopped being processed
properly: the strings where cut just before the first non-ASCII character and the application
would crash.  After spending a good deal of time on this, it turned out the problem was
in the code which converted the strings from multibyte (e.g. UTF-8) encoding which we were
given before, in MinGW-compiled code, to wide character encodings used by Windows. Interestingly, 
the problem was present even in code copied from Microsoft documentation for
Win32 APIs that were used to do the conversion and to obtain the command line arguments.

While looking for the cause, the code got modified to use `std::widestring` on Windows, in hope that
this would fix the issue.  Unfortunately, that wasn't the case - the conversion from multi-byte to wide
character set still failed in the same ways.  Eventually, I discovered that MSVC accepts the `/utf-8`
argument which causes it to switch both the source file encoding and the execution encoding (encoding
used for strings at the application run time) to UTF-8 - apparently what MinGW had done before.  This
fixed the issue in that the arguments passed to `main` can now be assumed to be `UTF-8` and most of the
code on both Windows and Unix can be shared again.  The wide string conversion **still** fails, though,
and for that reason the current code iteration uses only multi-byte `UTF-8` strings.

The infrastructure to handle different types of strings, depending on the platform, is still present, just in
case if we try to get wide character strings to work again some day.