Back to wide strings on Windows + magic encantations
The main fix here is the switch to `wmain` from `main` on
Windows.  The difference is that we're now passed parameters
encoded as Unicode wide strings, which makes them actually
keep the correct characters as specified on the application's
command line by the caller.

Also, in order to properly output wide strings to output streams,
applications need to set the correct mode for the output stream's
file handle (see `platform_setup`)