Bump to LLVM 18.1.8 (#31)
Changes: https://discourse.llvm.org/t/18-1-8-released/79725

There aren't any changes in this version that affect us directly.  However, we've noticed that
llc from 18.1.7 appears to be crashing with access violation on Windows and this bump is
an attempt to test whether the newer version of LLVM works better for us.