[ci] Add ManualValidation stage to create a GitHub release (#41)
A new stage has been added that will only run when someone manually
approves it. This stage will create a GitHub release with the relevant
artifacts from the build pipeline, making release management easier.

The tag used for the release will be generated based on the binutils
version and LLVM version.